beforefinal/!errormessage

+++ app.py
@@ -4,14 +4,17 @@ import bcrypt
 
 app = Flask(__name__)
 app.config['SECRET_KEY'] = '12rf1124g341h13'
-db = pymysql.connect(host='localhost', port=33061, user='root',passwd='hh237237!!',db='nextflexdb',charset='utf8')
+db = pymysql.connect(host='localhost', port=33061, user='root',passwd='*****',db='nextflexdb',charset='utf8')
 cursor=db.cursor()
 
  
 @app.route('/')
 @app.route('/index')
 def index():
-  return render_template('index.html')
+    sql = "SELECT * FROM movie ORDER BY NumberOfCopies DESC limit 3"
+    cursor.execute(sql)
+    topmovie=cursor.fetchall()
+    return render_template('index.html',topmovie=topmovie)
 
 @app.route('/movie_queue/<int:movie_id>')
 def movie_queue(movie_id):
@@ -27,9 +30,23 @@ def post_account():
     account_num = request.form['account_num']
     account_type = request.form['account_type']
     account_date = request.form['account_date']
-    print(user_id,account_num,account_type,account_date)
     sql = "INSERT INTO account(CustomerID,AccountNum,AccountType,PlanExpiredDate) VALUES (%s,%s,%s,DATE_ADD(NOW(), INTERVAL %s month));"
     cursor.execute(sql,(user_id,account_num,account_type,account_date))
+    db.commit()
+    return jsonify({'result': 'success'})
+
+@app.route('/returnmovie')
+def returnmovie():
+    user_id=session['user']
+    orderid=request.args.get('oid')
+    orderstar=request.args.get('ostar')
+    sql = "UPDATE Orders SET holding=0,Rating=%s WHERE OrderID=%s and CustomerID=%s"
+    cursor.execute(sql,(orderstar,orderid,user_id))
+    sql2 = "SELECT MovieID FROM Orders WHERE OrderID=%s"
+    cursor.execute(sql2,orderid)
+    orderedmovie=cursor.fetchone()
+    sql3 = "UPDATE nextflexdb.movie SET MovieRating= (SELECT AVG(Rating) FROM orders WHERE MovieID=%s) WHERE MovieID=%s"
+    cursor.execute(sql3,(orderedmovie,orderedmovie))
     db.commit()
     return jsonify({'result': 'success'})
 
@@ -37,12 +54,9 @@ def post_account():
 def viewmovie():
   user_id=session['user']
   orderid=request.args.get('oid')
-  sql = "SELECT * FROM Orders WHERE OrderID=%s"
-  cursor.execute(sql,orderid)
+  sql = "SELECT * FROM Orders WHERE OrderID=%s and CustomerID=%s and holding=1"
+  cursor.execute(sql,(orderid,user_id))
   ordercheck=cursor.fetchone()
-  #사용자 order match 검증
-  if(ordercheck[3]!=user_id):
-    return redirect('/mypage')
 
   sql2 = "SELECT * FROM movie WHERE MovieID=%s"
   cursor.execute(sql2,ordercheck[2])
@@ -54,17 +68,42 @@ def viewmovie():
 def makeorders():
   qid=request.args.get('qid')
   user_id=session['user']
-  sql = "SELECT * FROM moviequeue WHERE QueueId=%s"
-  cursor.execute(sql,qid)
+
+  #이미 order가 2개일 때 (return해야함) - unlimitedplan 
+  checksql = "SELECT COUNT(*) FROM orders WHERE CustomerId=%s and holding=1"
+  cursor.execute(checksql,user_id)
+  resultflag=cursor.fetchone()
+  if(resultflag[0]>=2):
+    return redirect('/mypage')
+    
+  #account에서 유효한 plan이 없을때
+  check2sql= "SELECT * FROM account WHERE CustomerID=%s and PlanExpiredDate>NOW();" 
+  cursor.execute(check2sql,user_id)
+  resultflag2=cursor.fetchall()
+  flag=0
+
+  #plans[4]-> movies column은 매달 1일 00시에 0으로 update
+  for plans in resultflag2:
+      if(plans[4]<2 or plans[3]=='Unlimited Plan'):
+          flag=1
+  if(flag==0):
+    return redirect('/mypage')
+
+  sql = "SELECT * FROM moviequeue WHERE QueueId=%s and CustomerID=%s"
+  cursor.execute(sql,(qid,user_id))
   resultflag=cursor.fetchone()
   if resultflag[2]==user_id:
     sql2 = "INSERT INTO orders(MovieID,CustomerID) SELECT MovieID,CustomerID FROM moviequeue WHERE QueueId=%s"
     cursor.execute(sql2,qid)
     sql3 = "DELETE FROM moviequeue WHERE QueueId=%s"
     cursor.execute(sql3,qid)
+    sql4 = "UPDATE account set movie=movie+1 WHERE CustomerID=%s;"
+    cursor.execute(sql4,user_id)
+    sql5 = "UPDATE movie set NumberOfCopies=NumberOfCopies+1 WHERE MovieID=%s;"
+    cursor.execute(sql5,resultflag[1])
     db.commit()
-  
   return redirect('/mypage')
+  
 
 
 @app.route('/movies')
@@ -89,6 +128,17 @@ def movies():
             movie_info=None
         return render_template('movies.html',movies=movie_info)
       elif(rtype=="actor"):
+        #두명 이상 검색
+        if("," in ktype):
+            ksplit=tuple(ktype.split(","))
+            knum=len(ktype.split(","))
+            sql = "SELECT * FROM movie left JOIN movieinfo_actor ON movie.MovieID = movieinfo_actor.Movie_info_ID left JOIN actor ON actor.ActorID = movieinfo_actor.Actor_info_ID where ActorName in %s GROUP BY MovieName HAVING COUNT(MovieName)=%s;"
+            rows_count = cursor.execute(sql,(ksplit,knum))
+            if rows_count > 0 :
+                movie_info = cursor.fetchall()
+            else:
+                movie_info=None
+            return render_template('movies.html',movies=movie_info)
         sql = "SELECT * FROM movie left JOIN movieinfo_actor ON movie.MovieID = movieinfo_actor.Movie_info_ID left JOIN actor ON actor.ActorID = movieinfo_actor.Actor_info_ID WHERE ActorName LIKE %s;"
         rows_count = cursor.execute(sql,(('%' + ktype + '%',)))
         if rows_count > 0 :
@@ -122,7 +172,7 @@ def mypage():
     cursor.execute(sql,user_id)
     user_info = cursor.fetchone()
 
-    sql2 = "SELECT * FROM account where CustomerID=%s"
+    sql2 = "SELECT * FROM account where CustomerID=%s and PlanExpiredDate > NOW()"
     cursor.execute(sql2,user_id)
     account_info = cursor.fetchall()
 
@@ -146,9 +196,10 @@ def logout():
 def about():
     if request.method == 'POST':
         login_info = request.form
+        print(login_info)
         EmailAddress = login_info['email']
         Passwd = login_info['passwd']
-        sql = "SELECT * FROM customer WHERE EmailAddress = %s"
+        sql = "SELECT * FROM customer WHERE EmailAddress=%s"
         rows_count = cursor.execute(sql,EmailAddress)
         if rows_count > 0 :
             user_info = cursor.fetchone()
@@ -173,12 +224,8 @@ def register():
         Passwd = bcrypt.hashpw(register_info['password1'].encode('utf-8'),bcrypt.gensalt())
         FirstName = register_info['Firstname']
         LastName = register_info['Lastname']
-        sql="""
-            INSERT INTO customer (EmailAddress,Passwd,FirstName,LastName)
-            VALUES (%s,%s,%s,%s);
-        """
+        sql="INSERT INTO customer (EmailAddress,Passwd,FirstName,LastName) VALUES (%s,%s,%s,%s);"
         cursor.execute(sql,(EmailAddress,Passwd,FirstName,LastName))
         db.commit()
-        db.close()
         return redirect('login')
     return render_template('signup.html')
